+++ trunk/src/AT_RDD_ExtendedTarget.c
@@ -125,8 +125,8 @@ void AT_RDD_ExtendedTarget_Gy( const long  n,
   float  max_electron_range_m = 0.0f;
   AT_max_electron_range_m(  n_tmp,
                 &E_MeV_u,
-                material_no,
-                er_model,
+                (int)material_no,
+                (int)er_model,
                 &max_electron_range_m);
 
   const double C_J_m               =  AT_RDD_Katz_C_J_m((double)electron_density_m3);
